Fix Arch repo update
Sort packages before adding them to the repo. Without this, updating the
repo can break if two versions of the same package are added in reverse
order, e.g. 5.5.10 before 5.5.9, as the 5.5.10 update will remove the
outdated 5.5.9 package. Since that, however, has already been provided
as command line option to the repo-add script, the script will try to
add it and then fail to find the package file, resulting in an error.